move auth component tests to folder
