+++ frontend/src/tests/auth_test/Register.test.js
@@ -1,4 +1,4 @@
-import { Register } from "../components";
+import { Register } from "../../components";
 
 test('Register', () => {
     const component = shallow(<Register />);
